Support attaching uprobes to targets inside containers
This commit adds a new executable called bpfman-ns which can be called
to attach uprobes and uretprobes inside containers.  This executable
changes to the mount namespace of the container, does the attach, and
then changes back to the original mount namespace.

The executable is spawned from the uprobe code when a container-pid
parameter is provided when loading a uprobe.

Other changes:
- renamed namespace to container_pid in bpfman proto and made it an int32

Signed-off-by: Andre Fredette <afredette@redhat.com>