Merge bpfctl into bpfman
This is the first phase of moving to a daemonless architecture.
This PR contains no changes to the existing logic of any of the
commands. It simply collapses the bpfman/bpfctl CLI into a single
binary.

- bpfman becomes "bpfman system service"
- all previous bpfctl commands are now bpfman commands

Some command renames have been performed to help with CLI
usability, as described in the design document.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>